+++ src/Commands/SWShortcuts.cpp
@@ -1,0 +1,1 @@
+#include "SWShortcuts.h"