+++ src/Ext/House/Body.h
@@ -35,6 +35,9 @@ public:
 		BuildingClass* Factory_NavyType;
 		BuildingClass* Factory_AircraftType;
 
+		int SuperWeaponButtonData[9];
+		int SuperWeaponButtonCount;
+
 		CDTimerClass AISuperWeaponDelayTimer;
 
 		//Read from INI
@@ -65,6 +68,8 @@ public:
 			, Factory_VehicleType { nullptr }
 			, Factory_NavyType { nullptr }
 			, Factory_AircraftType { nullptr }
+			, SuperWeaponButtonData { -1, -1, -1, -1, -1, -1, -1, -1, -1 }
+			, SuperWeaponButtonCount { 0 }
 			, AISuperWeaponDelayTimer {}
 			, RepairBaseNodes { false,false,false }
 			, RestrictedFactoryPlants {}