+++ src/Ext/Side/Body.cpp
@@ -41,6 +41,9 @@ void SideExt::ExtData::LoadFromINIFile(CCINIClass* pINI)
 	this->ToolTip_Background_Opacity.Read(exINI, pSection, "ToolTip.Background.Opacity");
 	this->ToolTip_Background_BlurSize.Read(exINI, pSection, "ToolTip.Background.BlurSize");
 	this->BriefingTheme = pINI->ReadTheme(pSection, "BriefingTheme", this->BriefingTheme);
+	this->SWSidebarBackground_TopPCX.Read(pINI, pSection, "SWSidebarBackground.TopPCX");
+	this->SWSidebarBackground_CenterPCX.Read(pINI, pSection, "SWSidebarBackground.CenterPCX");
+	this->SWSidebarBackground_BottomPCX.Read(pINI, pSection, "SWSidebarBackground.BottomPCX");
 }
 
 // =============================
@@ -70,6 +73,9 @@ void SideExt::ExtData::Serialize(T& Stm)
 		.Process(this->IngameScore_WinTheme)
 		.Process(this->IngameScore_LoseTheme)
 		.Process(this->BriefingTheme)
+		.Process(this->SWSidebarBackground_TopPCX)
+		.Process(this->SWSidebarBackground_CenterPCX)
+		.Process(this->SWSidebarBackground_BottomPCX)
 		;
 }