+++ src/Ext/Side/Body.h
@@ -36,6 +36,9 @@ public:
 		Nullable<int> ToolTip_Background_Opacity;
 		Nullable<float> ToolTip_Background_BlurSize;
 		Valueable<int> BriefingTheme;
+		PhobosPCXFile SWSidebarBackground_TopPCX;
+		PhobosPCXFile SWSidebarBackground_CenterPCX;
+		PhobosPCXFile SWSidebarBackground_BottomPCX;
 
 		ExtData(SideClass* OwnerObject) : Extension<SideClass>(OwnerObject)
 			, ArrayIndex { -1 }